Yoast SEO Update - Version 3.3.2

+++ plugins/wordpress-seo/admin/ajax.php
@@ -19,7 +19,7 @@ if ( ! defined( 'WPSEO_VERSION' ) ) {
  * @param array $results Results array for encoding.
  */
 function wpseo_ajax_json_echo_die( $results ) {
-	echo WPSEO_Utils::json_encode( $results );
+	echo wp_json_encode( $results );
 	die();
 }
 
@@ -43,6 +43,11 @@ function wpseo_set_option() {
 }
 
 add_action( 'wp_ajax_wpseo_set_option', 'wpseo_set_option' );
+
+/**
+ * Since 3.2 Notifications are dismissed in the Notification Center.
+ */
+add_action( 'wp_ajax_yoast_dismiss_notification', array( 'Yoast_Notification_Center', 'ajax_dismiss_notification' ) );
 
 /**
  * Function used to remove the admin notices for several purposes, dies on exit.
@@ -92,23 +97,35 @@ function wpseo_kill_blocking_files() {
 
 	check_ajax_referer( 'wpseo-blocking-files' );
 
-	$message = 'There were no files to delete.';
+	$message = 'success';
+	$errors  = array();
+
+	// Todo: Use WP_Filesystem, but not so easy to use in AJAX with credentials form still internal.
 	$options = get_option( 'wpseo' );
 	if ( is_array( $options['blocking_files'] ) && $options['blocking_files'] !== array() ) {
-		$message       = 'success';
-		$files_removed = 0;
-		foreach ( $options['blocking_files'] as $k => $file ) {
-			if ( ! @unlink( $file ) ) {
-				$message = __( 'Some files could not be removed. Please remove them via FTP.', 'wordpress-seo' );
+		foreach ( $options['blocking_files'] as $file ) {
+			if ( is_file( $file ) ) {
+				if ( ! @unlink( $file ) ) {
+					$errors[] = __(
+						sprintf( 'The file "%s" could not be removed. Please remove it via FTP.', $file ),
+						'wordpress-seo'
+					);
+				}
 			}
-			else {
-				unset( $options['blocking_files'][ $k ] );
-				$files_removed ++;
+
+			if ( is_dir( $file ) ) {
+				if ( ! @ rmdir( $file ) ) {
+					$errors[] = __(
+						sprintf( 'The directory "%s" could not be removed. Please remove it via FTP.', $file ),
+						'wordpress-seo'
+					);
+				}
 			}
 		}
-		if ( $files_removed > 0 ) {
-			update_option( 'wpseo', $options );
-		}
+	}
+
+	if ( $errors ) {
+		$message = implode( '<br />', $errors );
 	}
 
 	die( $message );
@@ -302,19 +319,6 @@ function wpseo_upsert_new( $what, $post_id, $new, $original ) {
 }
 
 /**
- * Create an export and return the URL
- */
-function wpseo_get_export() {
-
-	$include_taxonomy = ( filter_input( INPUT_POST, 'include_taxonomy' ) === 'true' );
-	$export           = new WPSEO_Export( $include_taxonomy );
-
-	wpseo_ajax_json_echo_die( $export->get_results() );
-}
-
-add_action( 'wp_ajax_wpseo_export', 'wpseo_get_export' );
-
-/**
  * Handles the posting of a new FB admin.
  */
 function wpseo_add_fb_admin() {
@@ -338,8 +342,12 @@ function ajax_get_keyword_usage() {
 	$post_id = filter_input( INPUT_POST, 'post_id' );
 	$keyword = filter_input( INPUT_POST, 'keyword' );
 
+	if ( ! current_user_can( 'edit_post', $post_id ) ) {
+		die( '-1' );
+	}
+
 	wp_die(
-		WPSEO_Utils::json_encode( WPSEO_Meta::keyword_usage( $keyword, $post_id ) )
+		wp_json_encode( WPSEO_Meta::keyword_usage( $keyword, $post_id ) )
 	);
 }
 
@@ -351,10 +359,25 @@ add_action( 'wp_ajax_get_focus_keyword_usage',  'ajax_get_keyword_usage' );
 function ajax_get_term_keyword_usage() {
 	$post_id = filter_input( INPUT_POST, 'post_id' );
 	$keyword = filter_input( INPUT_POST, 'keyword' );
-	$taxonomy = filter_input( INPUT_POST, 'taxonomy' );
+	$taxonomyName = filter_input( INPUT_POST, 'taxonomy' );
+
+	$taxonomy = get_taxonomy( $taxonomyName );
+
+	if ( ! $taxonomy ) {
+		wp_die( 0 );
+	}
+
+	if ( ! current_user_can( $taxonomy->cap->edit_terms ) ) {
+		wp_die( -1 );
+	}
+
+	$usage = WPSEO_Taxonomy_Meta::get_keyword_usage( $keyword, $post_id, $taxonomyName );
+
+	// Normalize the result so it it the same as the post keyword usage AJAX request.
+	$usage = $usage[ $keyword ];
 
 	wp_die(
-		WPSEO_Utils::json_encode( WPSEO_Taxonomy_Meta::get_keyword_usage( $keyword, $post_id, $taxonomy ) )
+		wp_json_encode( $usage )
 	);
 }
 
@@ -410,3 +433,16 @@ new WPSEO_Taxonomy_Columns();
 
 // Setting the notice for the recalculate the posts.
 new Yoast_Dismissable_Notice_Ajax( 'recalculate', Yoast_Dismissable_Notice_Ajax::FOR_SITE );
+
+/********************** DEPRECATED METHODS **********************/
+
+/**
+ * Create an export and return the URL
+ *
+ * @deprecated 3.3.2
+ */
+function wpseo_get_export() {
+	_deprecated_function( __METHOD__, 'WPSEO 3.3.2', 'This method is deprecated.' );
+
+	wpseo_ajax_json_echo_die( '' );
+}
